DEV: client side chta op code clean up
