+++ client/components/layout/StartScreen.tsx
@@ -1,7 +1,4 @@
-import useFetchConversation, {
-  useFetchChat,
-  useFetchChats
-} from "@/lib/hooks/useFetchConversation";
+import useFetchChats from "@/lib/hooks/useFetchChats";
 import { Image } from "@chakra-ui/react";
 import { Center, Alert, Text } from "@chakra-ui/react";
 import Spinner from "../general/Spinner";
@@ -28,8 +25,6 @@ type Props = {
 function StartScreen({ Child }: Props) {
   // initialize / fetch all queries for caching
   const { data, error } = useFetchChats();
-  // const { data, error } = useFetchConversation();
-  // useFetchChat();
 
   if (data) return Child;
