+++ client/lib/hooks/useFetchChats.tsx
@@ -1,0 +1,10 @@
+import chatOps from "@/graphql/operations/chat.ops";
+import { useQuery } from "@apollo/client";
+
+function useFetchChats() {
+  const { loading, data, error } = useQuery<ChatLean>(chatOps.Queries.getChats);
+
+  return { loading, data, error };
+}
+
+export default useFetchChats;
